// src/app/clients/Table.ts
export interface Table {
    contact_id: number;
    salutation: String;
    first_name: String;
    last_name: String;
    company_name: String;
    contact_type: String;
    client_number: String;
    interest_id: String;
    category_id: String;
    customer_category_id: String;
    level_id: String;
    catalogue_id: String;
    marital_status_id: String;
    marriage_anniversary_date: Date
    am_customer_id: String;
    approval_status_id: String;
    authorized_to_bid_id: String;
    email: String;
    phone: String;
    fax: String;
    mobile: String;
    other_phone: String;
    clients_vat_tin_no: String;
    aadhar_number: String;
    pan_no: string
    // date_created: Date
    // date_modified: Date
}